remove debugging idea
