+++ WordpressSync/debug.js
@@ -3,7 +3,7 @@ const { Values } = require('../../local-wordpress-to-github.settings.json');
 Object.keys(Values).forEach(x=>process.env[x]=Values[x]); //Load local settings file for testing
 
 //run the indexpage async
-const indexCode = require('./index.debug');
+const indexCode = require('./index');
 
 (async () => {
    return await indexCode();
