--- WordpressSync/endpoints.debug.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "./endpoints.schema.json",
-  "meta": {
-    "title": "endpoints config file",
-    "description": "endpoints config file"
-  },
-  "data": {
-    "projects": [
-      {
-        "name": "Drought",
-        "enabled": true,
-        "WordPressUrl": "https://live-drought-ca-gov.pantheonsite.io",
-        "HostUrl": {
-          "Production": "https://cagov.github.io/drought.ca.gov/",
-          "Development": "https://cagov.github.io/drought.ca.gov/"
-        },
-        "GitHubTarget": {
-          "Owner": "cagov",
-          "Repo": "drought.ca.gov",
-          "Branch": "media_test",
-          "PostPath": "wordpress/posts",
-          "PagePath": "wordpress/pages",
-          "SyncMedia": true,
-          "MediaPath": "wordpress/media"
-        }
-      }
-    ]
-  }
-}
